Major Merge
Pole-80
Pole-46
Pole-48
Pole-50
Pole-104
Onesin

Created first prototype for Setup room.

This includes, a working trash can with proper mesh, the table that will be used to place towers, a basic stick for now used to hit the table away. the Info board is also modeled but no info on it yet.  Added a couple generic torches since when i put the walls up.

also included a couple textures to give an idea of the theme/look